refactor: extract stateless watch utilities into watch-source-analysis.ts
Extract 12 stateless utility functions and types from watch.ts (2153→1914 lines)
into a new watch-source-analysis.ts module (296 lines). Functions moved:
- delayFileReadRetry, readSourceFileWithTransientEmptyRetry
- createExtensionMatcher, ExtensionMatcher interface
- countSourceLines, hashSourceContent
- resolveUnknownScanConcurrency, resolveDependentRetranspileConcurrency
- computeHotReloadLatencyStats
- takeInitialFileData, clearInitialFileDataCache, InitialFileData

All test imports updated. Build and lint pass clean.

Agent-Logs-Url: https://github.com/SimulatorLife/GMLoop/sessions/16138ea7-92fa-4ccb-85b9-398017432998
